refactor(Sheba): added constants & model, refactored file name & import path
