+++ example/example_sheba.dart
@@ -1,4 +1,5 @@
 import 'package:persian_tools/persian_tools.dart';
+import 'package:persian_tools/src/core/sheba/model.dart';
 
 void main() {
   var sheba = Sheba('IR820540102680020817909002');
