+++ test/test_sheba.dart
@@ -1,8 +1,9 @@
 import 'package:persian_tools/persian_tools.dart';
+import 'package:persian_tools/src/core/sheba/model.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('test Sheba/methods.dart', () {
+  group('test sheba.dart', () {
     test('test isValid getter', () {
       expect(Sheba('IR820540102680020817909002').isValid, equals(true));
       expect(Sheba('IR012345678901234567890123456789').isValid, equals(false));
